% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latinpalettes.R
\docType{data}
\name{latin_palettes}
\alias{latin_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 18.
}
\usage{
latin_palettes
}
\description{
Use \code{names(latin_palettes)} to view list of palette names.
Currently:  aventura, badbunny1, badbunny2, badgyal, beckyg, calle13, daddy1, daddy2,
don, ivyqueen, karolg, natti, nicky, ozuna, planb, rosalia, shakira, wyy.
}
\keyword{datasets}
